requirements.txt
google-adk>=1.0.0
python-dotenv>=1.0.0
redis>=5.0.0
requests>=2.31.0
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
reportlab
pydantic